% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvb.R
\name{fit.qvb}
\alias{fit.qvb}
\title{Fit the generalized q- von Bertalanffy growth funciton}
\usage{
fit.qvb(length, age, Lhat, r, q, tau, t0, method = "OLS", summary = FALSE)
}
\arguments{
\item{length}{length data}

\item{age}{age data}

\item{Lhat}{Initial parameter for L_hat; Growth scale factors}

\item{r}{Initial parameter for r; Growth exponents}

\item{q}{Initial parameter for q; Growth indeterminacy timing parameter}

\item{tau}{Initial parameter for tau; Maturation timing parameter}

\item{t0}{Initial parameter for t0; Theretical age at size zero}

\item{method}{Fitting method; select either "OLS" or "MLE"}

\item{summary}{if TRUE, summary is returned along with the estimated parameters}
}
\value{
list
}
\description{
Fit the generalized q- von Bertalanffy growth funciton
}
