% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loglike}
\alias{loglike}
\title{Calculate log likelihood}
\usage{
loglike(length, age, p = list(Linf, r, q, tau, t0), takelog = TRUE)
}
\arguments{
\item{length}{Length data}

\item{age}{Age data}

\item{p}{list
\describe{
\item{Linf}{Parameter L_inf}
\item{r}{Initial parameter for r}
\item{q}{Initial parameter for q \code{0 < q < 1} or \code{1 < q}}
\item{tau}{Initial parameter for tau}
\item{t0}{Initial parameter for t0}
}}

\item{takelog}{Set log = TRUE for log_likelihood}
}
\description{
Calculate log likelihood
}
