% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvb.R
\name{qvb}
\alias{qvb}
\title{Generalized q - von Bertalanffy growth funciton}
\usage{
qvb(age, Lhat, r, q, tau, t0)
}
\arguments{
\item{Lhat}{Initial parameter for L_hat; Growth scale factors}

\item{r}{Initial parameter for r; Growth exponents}

\item{q}{Initial parameter for q; Growth indeterminacy timing parameter}

\item{tau}{Initial parameter for tau; Maturation timing parameter}

\item{t0}{Initial parameter for t0; Theretical age at size zero}
}
\value{
vector of size at given age
}
\description{
Generalized q - von Bertalanffy growth funciton
}
\examples{
age <- seq(0, 10, 0.1)
qvb(age, Lhat=300, r=0.75, q=1.2, tau=2, t0= -0.01)
}
